requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
celery[redis]==5.3.6
redis==5.0.4
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
httpx==0.27.0
aiosmtplib==3.0.1
aioimaplib==1.1.0
aiogram>=3.4.1,<4
pydantic-settings==2.2.1
jinja2==3.1.4
python-multipart==0.0.9
slowapi==0.1.9
tenacity==8.3.0
email-validator==2.1.1
